More reliable error handling

+++ server/database/mongo.ts
@@ -4,7 +4,7 @@ import { v4 as uuid } from "uuid";
 import mongoose from "mongoose";
 
 const DB_PORT = 27017;
-const DB_URI = `mongodb://localhost:${DB_PORT}/testdb`;
+const DB_URI = `mongodb://localhost:${DB_PORT}/accountable`;
 
 /** Returns a fresh document ID that is virtually guaranteed not to have been used before. */
 export function newDocumentId(this: void): string {
@@ -14,8 +14,11 @@ export function newDocumentId(this: void): string {
 export async function fetchDbCollection<T extends AnyDataItem>(
 	ref: CollectionReference<T>
 ): Promise<Array<T>> {
-	const conn = mongoose.createConnection(DB_URI);
+	const mon = await mongoose.connect(DB_URI);
+	const conn = mon.connection;
 	const db = conn.db;
+	if (db === undefined) throw new EvalError("Database not initialized");
+
 	const collection = db.collection<T>(ref.id);
 	const query = collection.find({}).stream();
 
@@ -34,8 +37,11 @@ export async function fetchDbCollection<T extends AnyDataItem>(
 export async function fetchDbDoc<T extends AnyDataItem>(
 	ref: DocumentReference<T>
 ): Promise<T | null> {
-	const conn = mongoose.createConnection(DB_URI);
+	const mon = await mongoose.connect(DB_URI);
+	const conn = mon.connection;
 	const db = conn.db;
+	if (db === undefined) throw new EvalError("Database not initialized");
+
 	const collection = db.collection<AnyDataItem>(ref.parent.id);
 	return (await collection.findOne({ _id: ref.id })) as T;
 }
@@ -44,15 +50,21 @@ export async function upsertDbDoc<T extends AnyDataItem>(
 	ref: DocumentReference<T>,
 	data: T
 ): Promise<void> {
-	const conn = mongoose.createConnection(DB_URI);
+	const mon = await mongoose.connect(DB_URI);
+	const conn = mon.connection;
 	const db = conn.db;
+	if (db === undefined) throw new EvalError("Database not initialized");
+
 	const collection = db.collection(ref.parent.id);
 	await collection.replaceOne({ _id: ref.id }, data, { upsert: true });
 }
 
 export async function deleteDbDoc<T extends AnyDataItem>(ref: DocumentReference<T>): Promise<void> {
-	const conn = mongoose.createConnection(DB_URI);
+	const mon = await mongoose.connect(DB_URI);
+	const conn = mon.connection;
 	const db = conn.db;
+	if (db === undefined) throw new EvalError("Database not initialized");
+
 	const collection = db.collection(ref.parent.id);
 	await collection.deleteOne({ _id: ref.id });
 }
@@ -60,7 +72,10 @@ export async function deleteDbDoc<T extends AnyDataItem>(ref: DocumentReference<
 export async function deleteDbCollection<T extends AnyDataItem>(
 	ref: CollectionReference<T>
 ): Promise<void> {
-	const conn = mongoose.createConnection(DB_URI);
+	const mon = await mongoose.connect(DB_URI);
+	const conn = mon.connection;
 	const db = conn.db;
+	if (db === undefined) throw new EvalError("Database not initialized");
+
 	await db.dropCollection(ref.id);
 }
